Stage 1.7

+++ DeliveryWebApplication/Controllers/CouriersController.cs
@@ -8,123 +8,30 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using DeliveryWebApplication;
 using System.Globalization;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 
 namespace DeliveryWebApplication.Controllers
 {
-    public class CouriersController : Controller
+    public class CouriersController : MyController
     {
         private readonly DeliveryContext _context;
 
-        public CouriersController(DeliveryContext context)
-        {
-            _context = context;
-            Utils.SetCulture();
-        }
+        public CouriersController(DeliveryContext context, UserManager<User> userManager)
+            : base(userManager)
+            => _context = context;
 
         // GET: Couriers
         public async Task<IActionResult> Index()
         {
+            if (!CheckRoles(ADMIN)) return Forbid();
             return View(await _context.Couriers.ToListAsync());
-        }
-
-        // GET: Couriers/Details/5
-        public async Task<IActionResult> Details(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var courier = await _context.Couriers
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (courier == null)
-            {
-                return NotFound();
-            }
-
-            return View(courier);
-        }
-
-        // GET: Couriers/Create
-        public IActionResult Create()
-        {
-            return View();
-        }
-
-        // POST: Couriers/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Name,Phone")] Courier courier)
-        {
-            if (_context.Customers.Any(s => s.Phone == courier.Phone))
-                ModelState.AddModelError("Phone", "Кур'єр з таким номером телефона вже зареєестрований");
-            if (ModelState.IsValid)
-            {
-                _context.Add(courier);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            return View(courier);
-        }
-
-        // GET: Couriers/Edit/5
-        public async Task<IActionResult> Edit(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var courier = await _context.Couriers.FindAsync(id);
-            if (courier == null)
-            {
-                return NotFound();
-            }
-            return View(courier);
-        }
-
-        // POST: Couriers/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Name,Phone")] Courier courier)
-        {
-            if (id != courier.Id)
-            {
-                return NotFound();
-            }
-
-            if (_context.Customers.Any(s => s.Phone == courier.Phone && s.Id != courier.Id))
-                ModelState.AddModelError("Phone", "Кур'єр з таким номером телефона вже зареєестрований");
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(courier);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!CourierExists(courier.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            return View(courier);
         }
 
         // GET: Couriers/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (!CheckRoles(ADMIN)) return Forbid();
             if (id == null)
             {
                 return NotFound();
@@ -145,6 +52,7 @@ namespace DeliveryWebApplication.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!CheckRoles(ADMIN)) return Forbid();
             var courier = await _context.Couriers.FindAsync(id);
             _context.Couriers.Remove(courier);
             await _context.SaveChangesAsync();
